requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
pypdf>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
